Add generic personal number field to Checkout
